autoscroll to job running
