+++ webapp/views/App/JobMonitor/InnerJobs/InnerJob/index.js
@@ -1,0 +1,1 @@
+export { default } from './InnerJob'
